+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Support empty messages in `AnthropicTextProvider`.
+- Correctly send empty text content in `AnthropicTextProvider`.
 
 ## [0.4.0] - 2024-08-18