+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.1] - 2024-08-21
+
 ### Fixed
 
 - Allow tool result to be the last message in `AnthropicTextProvider`.
@@ -71,7 +73,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First public release.
 
-[unreleased]: https://gitlab.com/tozd/go/fun/-/compare/v0.5.0...main
+[unreleased]: https://gitlab.com/tozd/go/fun/-/compare/v0.5.1...main
+[0.5.1]: https://gitlab.com/tozd/go/fun/-/compare/v0.5.0...v0.5.1
 [0.5.0]: https://gitlab.com/tozd/go/fun/-/compare/v0.4.0...v0.5.0
 [0.4.0]: https://gitlab.com/tozd/go/fun/-/compare/v0.3.0...v0.4.0
 [0.3.0]: https://gitlab.com/tozd/go/fun/-/compare/v0.2.2...v0.3.0